requirements.txt
pyfiglet==0.8.post1
Pillow==8.3.1
